I kinda got dynamic input working.
I think I need to make the number input an option type, roughly my
desired behavior is if there is nothing in the input field then I don't
want anything to happen instead of the input field being overwritten.

I need to do error checking for strings, negative numbers, numbers that
are too small and so on. More research requred but I think this is it
for today.