Refactored model to just have workweight
I decided that having a model that contained more than the workweight
was a bit excessive. The model for now only needs to hold the workweight
because I think the program makes more sense if I generate the rest of
the webpage baised on that input.

Because the rest of the webpage is a flow of data from the one value
this makes more sense to me.